socrates_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socrates-teaching-env"
version = "1.0.0"
description = "RL environment for training LLMs to teach using the Socratic method"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0",
    "websockets>=11.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
training = [
    "trl>=0.7.0",
    "unsloth",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "wandb",
]
dev = [
    "pytest>=7.0",
    "httpx",
]

[tool.setuptools.packages.find]
include = ["socrates_env*", "server*", "training*"]
